feat(useHomePage.js): implement custom hook useHomePage for data fetching and pagination
